requirements.txt
antiorm==1.2.1
astroid==2.9.0
autopep8==1.6.0
click==8.0.3
db==0.1.1
db-sqlite3==0.0.1
Deprecated==1.2.13
Flask==2.0.2
flask-swagger-ui==3.36.0
importlib-metadata==4.8.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.12.2
redis==4.0.2
toml==0.10.2
typed-ast==1.5.1
typing-extensions==4.0.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
